+++ pyatlan/client/credential.py
@@ -29,7 +29,7 @@ class CredentialClient:
             )
         self._client = client
 
-    @validate_arguments()
+    @validate_arguments
     def get(self, guid: str) -> CredentialResponse:
         """
         Retrieves a credential by its unique identifier (GUID).
@@ -47,7 +47,7 @@ class CredentialClient:
             return raw_json
         return CredentialResponse(**raw_json)
 
-    @validate_arguments()
+    @validate_arguments
     def test(self, credential: Credential) -> CredentialTestResponse:
         """
         Tests the given credential by sending it to Atlan for validation.
@@ -55,34 +55,27 @@ class CredentialClient:
         :param credential: The credential to be tested.
         :type credential: A CredentialTestResponse instance.
         :returns: The response indicating the test result.
-        :raises InvalidRequestException: If the provided credential is invalid type.
+        :raises ValidationError: If the provided credential is invalid type.
         :raises AtlanException: On any error during API invocation.
         """
-        if not isinstance(credential, Credential):
-            raise ErrorCode.INVALID_PARAMETER_TYPE.exception_with_parameters(
-                "credential", "Credential"
-            )
         raw_json = self._client._call_api(TEST_CREDENTIAL, request_obj=credential)
         return CredentialTestResponse(**raw_json)
 
-    @validate_arguments()
+    @validate_arguments
     def test_and_update(self, credential: Credential) -> CredentialResponse:
         """
         Updates this credential in Atlan after first
         testing it to confirm its successful validation.
 
+        :param credential: The credential to be tested and updated.
         :returns: An updated CredentialResponse instance.
-        :raises InvalidRequestException: If the provided credential is invalid type.
+        :raises ValidationError: If the provided credential is invalid type.
         :raises InvalidRequestException: if the provided credentials
         cannot be validated successfully.
         :raises InvalidRequestException: If the provided credential
         does not have an ID.
         :raises AtlanException: on any error during API invocation.
         """
-        if not isinstance(credential, Credential):
-            raise ErrorCode.INVALID_PARAMETER_TYPE.exception_with_parameters(
-                "credential", "Credential"
-            )
         test_response = self.test(credential=credential)
         if not test_response.is_successful:
             raise ErrorCode.INVALID_CREDENTIALS.exception_with_parameters(